+++ swak/pt/misc.py
@@ -11,6 +11,8 @@ __all__ = [
     'NegativeBinomialFinalizer'
 ]
 
+# ToDo: Add Compile class
+
 
 def identity(tensor: Tensor, **_: Any) -> Tensor:
     """Simply pass through the argument and ignore keyword arguments.