+++ swak/__init__.py
@@ -1,1 +1,0 @@
-# ToDo: Make & use Literal types for Enums!